Try simplify Router
